+++ src/collective/eeafaceted/batchactions/tests/test_viewlets.py
@@ -71,8 +71,8 @@ class TestViewlets(BaseTestCase):
         viewlet = self._get_viewlet(self.eea_folder)
         self.assertEqual(
             viewlet.get_batch_actions(),
-            [{'name': 'transition-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 10}])
-
+            [{'name': 'transition-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 10},
+             {'name': 'labels-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 20}])
         # returned action names are traversable to get the form
         for action in viewlet.get_batch_actions():
             form = self.eea_folder.restrictedTraverse(action['name'])
@@ -104,13 +104,15 @@ class TestViewlets(BaseTestCase):
         viewlet = self._get_viewlet(folder)
         self.assertEqual(
             viewlet.get_batch_actions(),
-            [{'name': 'transition-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 10}])
+            [{'name': 'transition-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 10},
+             {'name': 'labels-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 20}])
 
         # mark with IBatchActionsSpecificMarker
         alsoProvides(folder, IBatchActionsSpecificMarker)
         self.assertEqual(
             viewlet.get_batch_actions(),
             [{'name': 'transition-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 10},
+             {'name': 'labels-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 20},
              {'name': 'testing-batch-action', 'button_with_icon': True, 'overlay': False, 'weight': 100}])
         # returned action names are traversable to get the form
         for action in viewlet.get_batch_actions():
@@ -121,4 +123,5 @@ class TestViewlets(BaseTestCase):
         viewlet = self._get_viewlet(self.eea_folder)
         self.assertEqual(
             viewlet.get_batch_actions(),
-            [{'name': 'transition-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 10}])
+            [{'name': 'transition-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 10},
+             {'name': 'labels-batch-action', 'button_with_icon': False, 'overlay': True, 'weight': 20}])